# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: LibroRed API
  description: REST API for LibroRed - Web Application for Book Lending Between Individuals
  contact:
    name: Team 13
    url: https://github.com/medinaymedia
    email: am.juradoc@alumnos.urjc.es
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: '1.0'
servers:
- url: https://localhost:8443
  description: Development server
...
paths:
...
  /api/loans/valid-borrowers:
    get:
      tags:
      - Loans
      summary: Get valid borrowers for authenticated user
      description: Get list of users who can borrow from the authenticated user
      operationId: getValidBorrowersForAuthUser
      responses:
        '200':
          description: Valid borrowers retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBasicDTO'
        '401':
          description: User not authenticated
  /api/v1/loans:
    get:
      tags:
      - Loans
      summary: Get all loans
      description: Retrieve a list of all loans
      operationId: getAllLoans
      responses:
        '200':
          description: Loans retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanDTO'
    post:
      tags:
      - Loans
      summary: Create a new loan
      description: Create a new loan entry
      operationId: createLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDTO'
        required: true
      responses:
        '201':
          description: Loan created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDTO'
        '400':
          description: Invalid loan data or business rule violation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDTO'
components:
  schemas:
...
    LoanDTO:
      type: object
      description: Loan data
      properties:
        id:
          type: integer
          format: int64
        book:
          $ref: '#/components/schemas/BookBasicDTO'
        lender:
          $ref: '#/components/schemas/UserBasicDTO'
        borrower:
          $ref: '#/components/schemas/UserBasicDTO'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - Active
          - Completed
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
    BookBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
...